% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_hertie.R
\docType{data}
\name{theme_hertie}
\alias{theme_hertie}
\title{Toy example ggplot theme for the Hertie School}
\format{
An object of class \code{theme} (inherits from \code{gg}) of length 93.
}
\usage{
theme_hertie
}
\description{
font_family: Georgia
base_size: 23
text colour: darkred
}
\details{
looks pretty crappy. Please improve <3
}
\keyword{datasets}
